// src/app/generated/cite.api/model/duty.ts
/**
 * Cite API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Evaluation } from './evaluation';
import { User } from './user';
import { Team } from './team';


export interface Duty { 
    dateCreated?: Date;
    dateModified?: Date | null;
    createdBy?: string;
    modifiedBy?: string | null;
    id?: string;
    evaluationId?: string;
    evaluation?: Evaluation;
    teamId?: string;
    team?: Team;
    name?: string | null;
    users?: Array<User> | null;
}